fix: null safety, strict warnings, comprehensive formatter tests
- guard toString(const char*) against nullptr
- add toString(std::nullptr_t) for C++11/14 compat
- enable -Wall -Wextra -Wpedantic (GCC/Clang) and /W4 (MSVC)
- require cxx_std_11 on INTERFACE target
- snprintf replaces ostringstream in toString(double/float)
- fix pre-epoch negative millisecond in timestamps
- normalize -$0.00 to $0.00 in currency format
- reject inf/nan as numeric in format specs
- remove dead heap fallback in snprintfDouble
- XML formatter uses string concat instead of ostringstream
- document flush() deadlock risk from sink write()
- add JSON escape tests (quotes, backslash, control chars, unicode)
- add XML sanitize/escape tests (empty key, digits, special chars)
- add tests for float, nullptr, null char*, autoFlush=false
- fix sign conversion in rate limit test
- regen single_include

+++ test/tests/test_rate_limiting.cpp
@@ -22,8 +22,8 @@ TEST_F(RateLimitingTest, EnforceRateLimit) {
     TestUtils::waitForFileContent("rate_limit_test_log.txt");
     std::string logContent = TestUtils::readLogFile("rate_limit_test_log.txt");
 
-    size_t messageCount = std::count(logContent.begin(), logContent.end(), '\n');
-    EXPECT_EQ(messageCount, 1000);
+    size_t messageCount = static_cast<size_t>(std::count(logContent.begin(), logContent.end(), '\n'));
+    EXPECT_EQ(messageCount, static_cast<size_t>(1000));
 }
 
 TEST_F(RateLimitingTest, ResetAfterRateLimit) {
